refactor: remove the project task-progress-bar feature
The project task-progress bar (the "% complete / N of M tasks" segmented
banner atop a project's task list) was previously disabled behind a dead
SHOW_TASK_PROGRESS_BAR flag. Remove the feature entirely — UI, the backend
aggregate endpoint/service, and the progress-only pure logic in @hezo/shared.

The same endpoint also fed the still-active onboarding phase banner and the
shared isLastRunFailed helper is used by the task list, so both are preserved:
the surviving banner data is repointed onto a lean, honestly-named endpoint.

- web: delete ProgressBar + SHOW_TASK_PROGRESS_BAR + project-status-badge;
  ProjectTaskListHeader now renders only the onboarding banner.
- rename hook use-task-progress-summary -> use-phase-banner and the query key
  tasksProgressSummary -> tasksPhaseBanner (websocket invalidation updated).
- server: replace getTaskProgressSummary with getProjectTaskListPhaseBanner and
  the route GET /tasks/progress-summary -> GET /tasks/phase-banner, returning
  just { phase_banner }.
- shared: drop buildTaskProgressSummary, deriveProjectStatus, ProjectStatus,
  the task-progress buckets and scope helpers; keep deriveProjectTaskListPhaseBanner
  and isLastRunFailed.
- tests: drop the progress-only suites; add a slim phase-banner endpoint test.
