+++ packages/web/src/hooks/use-websocket.ts
@@ -19,14 +19,14 @@ const TABLE_TO_QUERY_KEY: Record<
 > = {
 	tasks: (cid) => [
 		queryKeys.projects.tasks(cid),
-		queryKeys.projects.tasksProgressSummary(cid),
+		queryKeys.projects.tasksPhaseBanner(cid),
 		queryKeys.projects.all(),
 		queryKeys.projectIntakes(),
 	],
 	heartbeat_runs: (cid, row) => {
 		const keys: QueryKey[] = [
 			queryKeys.projects.tasks(cid),
-			queryKeys.projects.tasksProgressSummary(cid),
+			queryKeys.projects.tasksPhaseBanner(cid),
 		];
 		// A run starting/finishing flips the task's run-now availability (task_busy),
 		// so refresh that task's queued-wakeups (and their dispatch state).
