+++ packages/web/src/lib/query-keys.ts
@@ -65,7 +65,7 @@ export const queryKeys = {
 			'infinite',
 			filters,
 		],
-		tasksProgressSummary: (slug: string) => ['projects', slug, 'tasks', 'progress-summary'],
+		tasksPhaseBanner: (slug: string) => ['projects', slug, 'tasks', 'phase-banner'],
 		tasksResolve: (slug: string, key: string[]) => ['projects', slug, 'tasks', 'resolve', key],
 		task: (slug: string, taskId: string) => ['projects', slug, 'tasks', taskId],
 		taskComments: (slug: string, taskId: string) => ['projects', slug, 'tasks', taskId, 'comments'],
